Add tooltips to history markers (#21)
Gives the ability to view exact values for
a certain location in history.  Still need
to add the ability to choose how many in history
to store as well as figure out a way to pop history
markers to front on pause so the tooltips can work when
behind the sensor.